release on 2016-05-18_11-38-54

+++ src/lagrangian/cfdemParticle/etc/functions.sh
@@ -153,7 +153,7 @@ compileSolver()
     
     # compile parallel?
     if [[ $parallel == "true" ]]; then
-        wmake 2>&1 | tee -a $logpath/$logfileName &
+        wmake 2>&1 | tee -a $logpath/$logfileName #&
     else
         wmake 2>&1 | tee -a $logpath/$logfileName
     fi
@@ -224,10 +224,8 @@ compileLMPlib()
     headerText="$3"
     makeFileName="$4"
     libraryPath="$5"
+    compilationModeSwitch="$6"
     #--------------------------------------------------------------------------------#
-
-    #- clean up old log file
-    rm $logpath/$logfileName
 
     #- change path
     if [ -d "$libraryPath" ]; then
@@ -238,29 +236,64 @@ compileLMPlib()
         read
     fi
 
-    #- header
-    echo 2>&1 | tee -a $logpath/$logfileName
-    echo "//   $headerText   //" 2>&1 | tee -a $logpath/$logfileName
-    echo 2>&1 | tee -a $logpath/$logfileName
-
-    #- write path
-    pwd 2>&1 | tee -a $logpath/$logfileName
-    echo 2>&1 | tee -a $logpath/$logfileName
-
-    if [[ $makeFileName == "Makefile.Install" ]]; then
-        echo "using Install.sh"
-        bash Install.sh 0 2>&1 | tee -a $logpath/$logfileName
-        bash Install.sh 1 2>&1 | tee -a $logpath/$logfileName
+    #Just check if library is there and and abort if not
+    if [ $compilationModeSwitch == "false" ]; then
+          if [ -d "$libraryPath" ]; then
+                    echo "lib path $libraryPath EXISTS!"
+                    cd $libraryPath
+                    if [ -e *.a ]; then
+                        echo "... and contains the following libraries: "
+                        ls $libraryPath/*.a
+                        echo "Congratulations! Check passed! "
+                    else
+                        echo ""
+                        echo "ERROR!!"
+                        echo "... could not find Library!"
+                        echo "... it should contain a *.a file to be linked to the final application."
+                        echo "You need to ensure all libaries in this path are compiled."
+                        echo "$libraryPath"
+                        echo "are compiled. Therefore, you may want to use an appropriate script that compiles these libraries first. (e.g. cfdemCompLIGlibs)"
+                        read
+                    fi
+          fi
     else
-        #- clean up
-        echo "make clean" 2>&1 | tee -a $logpath/$logfileName
+        if [ $compilationModeSwitch == "noClean" ]; then
+            echo "compileLMPlib will skip the cleaning step!"
+            echo ""
+        else
+            #Clean and then compile library
+            #- clean up old log file
+            rm $logpath/$logfileName
+        fi
+        #- header
         echo 2>&1 | tee -a $logpath/$logfileName
-        make -f $makeFileName clean 2>&1 | tee -a $logpath/$logfileName
+        echo "//   $headerText   //" 2>&1 | tee -a $logpath/$logfileName
+        echo 2>&1 | tee -a $logpath/$logfileName
 
-        #- compile
-        echo "make" 2>&1 | tee -a $logpath/$logfileName
+        #- write path
+        pwd 2>&1 | tee -a $logpath/$logfileName
         echo 2>&1 | tee -a $logpath/$logfileName
-        make -f $makeFileName 2>&1 | tee -a $logpath/$logfileName
+
+        if [[ $makeFileName == "Makefile.Install" ]]; then
+            echo "using Install.sh"
+            bash Install.sh 0 2>&1 | tee -a $logpath/$logfileName
+            bash Install.sh 1 2>&1 | tee -a $logpath/$logfileName
+        else
+            if [ $compilationModeSwitch == "noClean" ]; then
+                echo "compileLMPlib will skip the cleaning step!"
+                echo ""
+            else
+                #- clean up
+                echo "make clean" 2>&1 | tee -a $logpath/$logfileName
+                echo 2>&1 | tee -a $logpath/$logfileName
+                make -f $makeFileName clean 2>&1 | tee -a $logpath/$logfileName
+            fi
+
+            #- compile
+            echo "make" 2>&1 | tee -a $logpath/$logfileName
+            echo 2>&1 | tee -a $logpath/$logfileName
+            make -f $makeFileName 2>&1 | tee -a $logpath/$logfileName
+        fi
     fi
 }
 #==================================#
@@ -278,21 +311,21 @@ cleanCFDEM()
     #cleaning libraries
     whitelist="$CFDEM_SRC_DIR/lagrangian/cfdemParticle/etc/library-list.txt"
     echo ""
-    echo "Please provide the libraries to be cleaned in the $CWD/$whitelist file."
+    echo "Please provide the libraries to be cleaned in the $whitelist file."
 
-    if [ ! -f "$CWD/$whitelist" ];then
+    if [ ! -f "$whitelist" ];then
         echo "$whitelist does not exist in $CWD. Nothing will be done."
         NLINES=0
         COUNT=0
     else
-        NLINES=`wc -l < $CWD/$whitelist`
+        NLINES=`wc -l < $whitelist`
         COUNT=0
     fi
 
     while [ $COUNT -lt $NLINES ]
     do
             let COUNT++  
-            LINE=`head -n $COUNT $CWD/$whitelist | tail -1`
+            LINE=`head -n $COUNT $whitelist | tail -1`
   
             # white lines
             if [[ "$LINE" == "" ]]; then
@@ -336,21 +369,21 @@ cleanCFDEM()
     #cleaning solvers
     whitelist="$CFDEM_SRC_DIR/lagrangian/cfdemParticle/etc/solver-list.txt"
     echo ""
-    echo "Please provide the solvers to be cleaned in the $CWD/$whitelist file."
+    echo "Please provide the solvers to be cleaned in the $whitelist file."
 
-    if [ ! -f "$CWD/$whitelist" ];then
+    if [ ! -f "$whitelist" ];then
         echo "$whitelist does not exist in $CWD. Nothing will be done."
         NLINES=0
         COUNT=0
     else
-        NLINES=`wc -l < $CWD/$whitelist`
+        NLINES=`wc -l < $whitelist`
         COUNT=0
     fi
 
     while [ $COUNT -lt $NLINES ]
     do
             let COUNT++  
-            LINE=`head -n $COUNT $CWD/$whitelist | tail -1`
+            LINE=`head -n $COUNT $whitelist | tail -1`
   
             # white lines
             if [[ "$LINE" == "" ]]; then
@@ -376,6 +409,10 @@ cleanCFDEM()
             fi
             wclean    
     done
+
+    #**********************************************
+    #cleaning logs
+    rm $CFDEM_SRC_DIR/lagrangian/cfdemParticle/etc/log/log_*
 }
 #==================================#
 
@@ -419,6 +456,9 @@ DEMrun()
     elif [ $debugMode == "strict" ]; then
         #debugMode="valgrind --leak-check=full -v --trace-children=yes --track-origins=yes" 
         debugMode="valgrind --tool=memcheck --track-origins=yes --leak-check=yes --show-reachable=yes --num-callers=20 --track-fds=yes"  
+    elif [ $debugMode == "strictXML" ]; then
+        #debugMode="valgrind --leak-check=full -v --trace-children=yes --track-origins=yes" 
+        debugMode="valgrind --tool=memcheck --leak-check=yes --leak-check-heuristics=stdstring,length64,newarray,multipleinheritance --show-reachable=no --num-callers=20 --track-fds=yes --xml=yes --xml-file=valgrind.xml"  
     else
         debugMode=""
     fi
@@ -467,7 +507,10 @@ parDEMrun()
         debugMode="valgrind"
     elif [ $debugMode == "strict" ]; then
         #debugMode="valgrind --leak-check=full -v --trace-children=yes --track-origins=yes" 
-        debugMode="valgrind --tool=memcheck --leak-check=yes --show-reachable=yes --num-callers=20 --track-fds=yes"  
+        debugMode="valgrind --tool=memcheck --track-origins=yes --leak-check=yes --show-reachable=yes --num-callers=20 --track-fds=yes"  
+    elif [ $debugMode == "strictXML" ]; then
+        #debugMode="valgrind --leak-check=full -v --trace-children=yes --track-origins=yes" 
+        debugMode="valgrind --tool=memcheck --leak-check=yes --leak-check-heuristics=stdstring,length64,newarray,multipleinheritance --show-reachable=no --num-callers=20 --track-fds=yes --xml=yes --xml-file=valgrind.xml"  
     else
         debugMode=""
     fi
@@ -625,15 +668,19 @@ parCFDDEMrun()
     machineFileName="$7"
     debugMode="$8"
     reconstuctCase="$9"
-    cleanCase="$10"
+    decomposeCase=${10}
     #--------------------------------------------------------------------------------#
 
     if [ $debugMode == "on" ]; then
         debugMode="valgrind"
     elif [ $debugMode == "strict" ]; then
         #debugMode="valgrind --leak-check=full -v --trace-children=yes --track-origins=yes" 
-        debugMode="valgrind --tool=memcheck --leak-check=yes --show-reachable=yes --num-callers=20 --track-fds=yes"
+        debugMode="valgrind --tool=memcheck --track-origins=yes --leak-check=yes --show-reachable=yes --num-callers=20 --track-fds=yes"
+    elif [ $debugMode == "strictXML" ]; then
+        #debugMode="valgrind --leak-check=full -v --trace-children=yes --track-origins=yes" 
+        debugMode="valgrind --tool=memcheck --leak-check=yes --leak-check-heuristics=stdstring,length64,newarray,multipleinheritance --show-reachable=no --num-callers=20 --track-fds=yes --xml=yes --xml-file=valgrind.xml"  
     elif [ $debugMode == "profile" ]; then
+        # make sure you did hpcstruct before
         debugMode="hpcrun"
         rm -r $casePath/CFD/hpctoolkit-$solverName-measurements
     else
@@ -646,11 +693,16 @@ parCFDDEMrun()
     #- change path
     cd $casePath/CFD
 
-    #- remove old data
-    rm -rf processor*
-
     #- decompose case
-    decomposePar
+    if [[ $decomposeCase == "false" ]]; then   
+        echo "Not decomposing case."
+    else
+        echo "Decomposing case."
+        #- remove old data
+        rm -rf processor*
+
+        decomposePar
+    fi
 
     #- make proc dirs visible
     count=0
